Reorganizando os modelos
Modelos reorganizados e adição do Twitter Boostrap para estilzar as
views. Modelos reorganizados: Curso, Aula.

+++ db/migrate/20121009035711_create_matriculas.rb
@@ -3,7 +3,7 @@ class CreateMatriculas < ActiveRecord::Migration
     create_table :matriculas do |t|
       t.references :aluno
       t.references :curso
-      t.double :nota
+      t.float :nota
 
       t.timestamps
     end
